backend_macos/requirements.txt
fastapi>=0.115,<1.0
uvicorn[standard]>=0.30,<1.0
pyobjc-framework-Vision>=12.1
pyobjc-framework-Quartz>=12.1
pyobjc-framework-Cocoa>=12.1
